bbe.conf: Fix do_image_wic dependencies
Depending on u-boot-bbe:do_build can have some strange effects when the
rm_work class is used. The extend_recipe_sysroot function finds the
following dependencies for do_image_wic:

    do_image_wic
        -> u-boot-bbe:do_build
        -> u-boot-bbe:do_rm_work_all
        -> glibc-initial:do_rm_work
        -> glibc-initial:do_populate_sysroot

This results in an attempt to install both glibc and glibc-initial into
the recipe sysroot which causes the build to fail.

Instead we should be depending on u-boot-bbe:do_deploy as we just need
the u-boot artifacts to be present in the deploy directory.

Signed-off-by: Paul Barker <paul.barker@sancloud.co.uk>