Switch back to u-boot script for booting
It's not possible to automatically detect the partition we're booting
from and set the kernel command line appropriately when using an
extlinux.conf file. This means the rootfs image can only support one
boot medium at a time (SD card) and would need to be modified to boot
from an alternative medium (eMMC, USB, etc).

To fix this we need to switch back to using a u-boot script which can
run the `part uuid ...` command. The script is updated as the
`loadimage` and `loadfdt` variables we previously used no longer exist
in u-boot v2021.10.

Ref: BSP-70
Signed-off-by: Paul Barker <paul.barker@sancloud.com>