linux-bbe: Drop unused cmem & ti-uio features
After reviewing the meta-ti layer it was seen that these features are
not enabled in any of our builds. Keeping support for these features
requires us to carry patch dtsi files in our layer which have ambiguous
licensing so let's drop this support for now to simplify things and help
us to achieve REUSE compliance. We can add this support back in the
future if we need to.

Signed-off-by: Paul Barker <paul.barker@sancloud.com>
Ref: BSP-73